refactor: rename the cohortColums
